requirements.txt
SpeechRecognition==3.10.0
pydub==0.25.1
googletrans==4.0.0-rc1
openai-whisper==20231117
transformers>=4.30.0
torch>=2.0.0
torchaudio>=2.0.0
nemo_toolkit[asr]>=2.4.0
librosa>=0.11.0
soundfile>=0.13.0
torch>=2.0.0
torchaudio>=2.0.0
transformers>=4.30.0
psutil>=5.9.0